Database update
* optimized code
* increased stability
* logs current Database version
* removed need for calofornium properties file
* better validation
